local project update

+++ project/uni_scalar.m
@@ -14,7 +14,7 @@ function D = uni_scalar(iSignal, inA, desiredL)
         end
         
     else %signal is 2-D
-        D=uint8(zeros(m,n));
+        D=zeros(m,n);
         for j=1:m
             for k=1:n
                 roundVal = floor( (abs(iSignal(j,k))/step) + (1/2) );
